Refactor documentation into its own file
