+++ stub-generator/.gitignore
@@ -1,1 +1,1 @@
-matrix_analytics_docgen/__pycache__/
+**/__pycache__/
